+++ src/test/java/net/adeptropolis/frogspawn/clustering/postprocessing/postprocessors/AffiliationGuardingPostprocessorTest.java
@@ -39,7 +39,7 @@ public class AffiliationGuardingPostprocessorTest {
   @Before
   public void setUp() {
     metric = new DefaultAffiliationMetric();
-    graph = new SparseGraphBuilder(0)
+    graph = new SparseGraphBuilder()
             .add(0, 1, 1)
             .add(0, 2, 1)
             .add(2, 3, 1)